+++ src/lib/utils/formatter/token.ts
@@ -16,23 +16,26 @@ export const formatDecimal =
     delimiter: boolean;
   }) =>
   (n: BigSource, fallbackValue: string): string => {
-    const num = big(
-      big(n)
-        .mul(10 ** decimalPoints)
-        .toFixed()
-        .split(".")[0]
-    )
-      .div(10 ** decimalPoints)
-      .toFixed(decimalPoints);
-    if (num === "NaN") return fallbackValue;
+    try {
+      const num = big(
+        big(n)
+          .mul(10 ** decimalPoints)
+          .toFixed()
+          .split(".")[0]
+      )
+        .div(10 ** decimalPoints)
+        .toFixed(decimalPoints);
 
-    const [i, d] = num.split(".");
-    const thousands = /\B(?=(\d{3})+(?!\d))/g;
+      const [i, d] = num.split(".");
+      const thousands = /\B(?=(\d{3})+(?!\d))/g;
 
-    const ii = delimiter ? i.replace(thousands, ",") : i;
-    const dd = d ? `.${d}` : "";
+      const ii = delimiter ? i.replace(thousands, ",") : i;
+      const dd = d ? `.${d}` : "";
 
-    return (ii === "0" && num[0] === "-" ? "-" : "") + ii + dd;
+      return (ii === "0" && num[0] === "-" ? "-" : "") + ii + dd;
+    } catch (e) {
+      return fallbackValue;
+    }
   };
 
 export const d2Formatter = formatDecimal({ decimalPoints: 2, delimiter: true });
@@ -42,7 +45,13 @@ export const toToken = (
   uAmount: U<Token<BigSource>>,
   precision: number
   // TODO: try toFixed(precision) here after having unit test?
-): Token<Big> => big(uAmount).div(big(10).pow(precision)) as Token<Big>;
+): Token<Big> => {
+  try {
+    return big(uAmount).div(big(10).pow(precision)) as Token<Big>;
+  } catch {
+    return big(0) as Token<Big>;
+  }
+};
 
 /**
  * @remarks
@@ -75,20 +84,24 @@ export const formatUTokenWithPrecision = (
  * <$0.000001 -> <$0.000001
  */
 export const formatPrice = (value: USD<BigSource>): string => {
-  const price = big(value);
-  const lowestThreshold = 0.000001;
+  try {
+    const price = big(value);
+    const lowestThreshold = 0.000001;
 
-  const d2 = d2Formatter(price, "0.00");
-  const d6 = d6Formatter(price, "0.00");
+    const d2 = d2Formatter(price, "0.00");
+    const d6 = d6Formatter(price, "0.00");
 
-  if (price.eq(0) || price.gte(1)) {
-    return `$${d2}`;
+    if (price.eq(0) || price.gte(1)) {
+      return `$${d2}`;
+    }
+
+    if (price.lt(lowestThreshold)) {
+      return `<$${lowestThreshold}`;
+    }
+    return `$${d6}`;
+  } catch {
+    return "$0.00";
   }
-
-  if (price.lt(lowestThreshold)) {
-    return `<$${lowestThreshold}`;
-  }
-  return `$${d6}`;
 };
 
 export const formatInteger = (n: BigSource): string => {